refactor(dask): remove pre_execute
