+++ ibis/backends/dask/__init__.py
@@ -30,7 +30,7 @@ class Backend(BasePandasBackend):
 
     def do_connect(self, dictionary):
         # register dispatchers
-        from . import udf  # noqa: F401
+        from . import execution  # noqa: F401
 
         super().do_connect(dictionary)
