+++ ibis/backends/dask/tests/test_core.py
@@ -17,10 +17,10 @@ from dask.dataframe.utils import tm  # noqa: E402
 from ibis.backends.dask import Backend  # noqa: E402
 
 from ..core import execute, is_computable_input  # noqa: E402
-from ..dispatch import execute_node, pre_execute  # noqa: E402
+from ..dispatch import execute_node  # noqa: E402
 
 
-@pytest.mark.parametrize('func', [execute_node, pre_execute])
+@pytest.mark.parametrize('func', [execute_node])
 def test_no_execute_ambiguities(func):
     assert not ambiguities(func.funcs)
 
@@ -39,26 +39,6 @@ def test_from_dataframe(dataframe, ibis_table, core_client):
     t = ibis.dask.from_dataframe(dataframe, name='foo', client=client)
     expected = ibis_table.execute()
     tm.assert_frame_equal(result, expected)
-
-
-def test_pre_execute_basic():
-    """
-    Test that pre_execute has intercepted execution and provided its own
-    scope dict
-    """
-
-    @pre_execute.register(ops.Add)
-    def pre_execute_test(op, *clients, scope=None, **kwargs):
-        return Scope({op: 4}, None)
-
-    one = ibis.literal(1)
-    expr = one + one
-    result = execute(expr)
-    assert result == 4
-
-    del pre_execute.funcs[(ops.Add,)]
-    pre_execute.reorder()
-    pre_execute._cache.clear()
 
 
 def test_execute_parameter_only():
